+++ Collector/include/data_tuple.h
@@ -39,9 +39,9 @@ public:
 		const stk_charge &extracted_stk_charge,
 		const bgo_classifiers &classifier,
 		const trigger_info &evt_trigger,
-		const std::vector<double> &nud_adc,
-		const double nud_total_adc,
-		const double nud_max_adc,
+		const std::vector<int> &nud_adc,
+		const int nud_total_adc,
+		const int nud_max_adc,
 		const int nud_max_channel_id);
 	void Reset();
